ENH: refactor
